CameraOCRApp/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
PyQt5==5.15.9
numpy==2.2.6
pyodbc==4.0.39
requests==2.32.5